+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed the _Postgres_ image from `postgres:15` to `postgres:15-alpine` in the `docker-compose` files
 - Introduced `extends` in the `docker-compose` files
 - Improved the language localization for German (`de`)
+- Removed some uses of `lodash` helpers
 
 ## 2.132.0 - 2024-12-30